+++ netbox_ssl/utils/__init__.py
@@ -1,5 +1,6 @@
 from .compliance_checker import CheckResult, ComplianceChecker
 from .csr_parser import CSRParseError, CSRParser
+from .export import CertificateExporter, ExportFormatChoices
 from .parser import CertificateParseError, CertificateParser, PrivateKeyDetectedError
 
 __all__ = [
@@ -10,4 +11,6 @@ __all__ = [
     "CSRParseError",
     "ComplianceChecker",
     "CheckResult",
+    "CertificateExporter",
+    "ExportFormatChoices",
 ]